Hierarchical model MCMC

Run the hierarchical model with MCMC.

hb_mcmc_hierarchical(
  data,
  response = "response",
  study = "study",
  study_reference = max(data[[study]]),
  group = "group",
  group_reference = min(data[[group]]),
  patient = "patient",
  covariates = grep("^covariate", colnames(data), value = TRUE),
  s_delta = 30,
  s_beta = 30,
  s_sigma = 30,
  s_mu = 30,
  s_tau = 30,
  n_chains = 4,
  n_adapt = 2000,
  n_warmup = 4000,
  n_iterations = 20000,
  quiet = TRUE
)

Arguments

data

Tidy data frame with one row per patient, indicator columns for the response variable, study, group, and patient, and covariates. All columns must be atomic vectors (e.g. not lists). The data for the mixture and simple models should have just one study, and the others should have data from more than one study. The simple model can be used to get the historical data components of m_omega and s_omega for the mixture model.

response

Character of length 1, name of the column in data with the response/outcome variable. data[[response]] must be a continuous variable, and it should be the change from baseline of a clinical endpoint of interest, as opposed to just the raw response. Treatment differences are computed directly from this scale, please supply change from baseline unless you are absolutely certain that treatment differences computed directly from this quantity are clinically meaningful.

study

Character of length 1, name of the column in data with the study ID.

study_reference

Atomic of length 1, element of the study column that indicates the current study. (The other studies are historical studies.)

group

Character of length 1, name of the column in data with the group ID.

group_reference

Atomic of length 1, element of the group column that indicates the control group. (The other groups may be treatment groups.)

patient

Character of length 1, name of the column in data with the patient ID.

covariates

Character vector of column names in data with the columns with baseline covariates. These can be continuous, categorical, or binary. Regardless, historicalborrow derives the appropriate model matrix.

s_delta

Numeric of length 1, prior standard deviation of the study-by-group effect parameters delta.

s_beta

Numeric of length 1, prior standard deviation of the fixed effects beta.

s_sigma

Numeric of length 1, prior upper bound of the residual standard deviations.

s_mu

Positive numeric of length 1, hyperparameter: prior standard deviation of the mean mu of the study-level control means alpha.

s_tau

Positive numeric of length 1, hyperparameter: uniform prior upper bound of the standard deviation tau of the study-level control means alpha.

n_chains

Number of MCMC chains to run.

n_adapt

Number of adaptation iterations to run.

n_warmup

Number of warmup iterations per chain to run.

n_iterations

Number of saved MCMC iterations per chain to run.

quiet

Logical of length 1, TRUE to suppress R console output.

Value

A tidy data frame of parameter samples from the posterior distribution. Columns .chain, .iteration, and .draw have the meanings documented in the posterior package.

See also

Examples

data <- hb_sim_hierarchical(n_continuous = 2)$data
hb_mcmc_hierarchical(
  data,
  n_chains = 1,
  n_adapt = 100,
  n_warmup = 50,
  n_iterations = 50
)
#> # A tibble: 50 × 28
#>    `alpha[1]` `alpha[2]` alpha…¹ alpha…² alpha…³ beta[…⁴ beta[…⁵ beta[…⁶ beta[…⁷
#>         <dbl>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1      -1.67      -1.72   -1.68   -1.71   -1.70    1.33 -0.255   -0.273   0.386
#>  2      -1.71      -1.73   -1.68   -1.74   -1.70    1.52 -0.306   -0.143   0.349
#>  3      -1.67      -1.70   -1.71   -1.72   -1.70    1.51 -0.0316  -0.163   0.325
#>  4      -1.73      -1.74   -1.70   -1.73   -1.69    1.47 -0.0929  -0.201   0.279
#>  5      -1.71      -1.69   -1.68   -1.72   -1.73    1.58 -0.337   -0.318   0.268
#>  6      -1.70      -1.74   -1.74   -1.70   -1.74    1.58 -0.107   -0.183   0.343
#>  7      -1.68      -1.72   -1.64   -1.70   -1.69    1.43 -0.260   -0.227   0.334
#>  8      -1.74      -1.78   -1.69   -1.72   -1.65    1.50 -0.323   -0.179   0.364
#>  9      -1.69      -1.74   -1.64   -1.70   -1.68    1.44 -0.369   -0.214   0.397
#> 10      -1.60      -1.77   -1.67   -1.71   -1.69    1.39 -0.216   -0.168   0.269
#> # … with 40 more rows, 19 more variables: `beta[5]` <dbl>, `beta[6]` <dbl>,
#> #   `beta[7]` <dbl>, `beta[8]` <dbl>, `beta[9]` <dbl>, `beta[10]` <dbl>,
#> #   `delta[1]` <dbl>, `delta[2]` <dbl>, mu <dbl>, precision_ratio <dbl>,
#> #   `sigma[1]` <dbl>, `sigma[2]` <dbl>, `sigma[3]` <dbl>, `sigma[4]` <dbl>,
#> #   `sigma[5]` <dbl>, tau <dbl>, .chain <int>, .iteration <int>, .draw <int>,
#> #   and abbreviated variable names ¹​`alpha[3]`, ²​`alpha[4]`, ³​`alpha[5]`,
#> #   ⁴​`beta[1]`, ⁵​`beta[2]`, ⁶​`beta[3]`, ⁷​`beta[4]`