Mixture model MCMC hyperparameters

Run a simple model separately on each historical study control group and compute hyperparameters for hb_mcmc_mixture().

hb_mcmc_mixture_hyperparameters(
  data,
  response = "response",
  study = "study",
  study_reference = max(data[[study]]),
  group = "group",
  group_reference = min(data[[group]]),
  patient = "patient",
  m_mu = 0,
  s_mu = 30,
  s_sigma = 30,
  m_omega_current = 0,
  s_omega_current = 30,
  n_chains = 4,
  n_adapt = 2000,
  n_warmup = 4000,
  n_iterations = 20000,
  quiet = TRUE
)

Arguments

data

Tidy data frame with one row per patient, indicator columns for the response variable, study, group, and patient, and covariates. All columns must be atomic vectors (e.g. not lists). The data for the mixture and simple models should have just one study, and the others should have data from more than one study. The simple model can be used to get the historical data components of m_omega and s_omega for the mixture model.

response

Character of length 1, name of the column in data with the response/outcome variable. data[[response]] must be a continuous variable, and it should be the change from baseline of a clinical endpoint of interest, as opposed to just the raw response. Treatment differences are computed directly from this scale, please supply change from baseline unless you are absolutely certain that treatment differences computed directly from this quantity are clinically meaningful.

study

Character of length 1, name of the column in data with the study ID.

study_reference

Atomic of length 1, element of the study column that indicates the current study. (The other studies are historical studies.)

group

Character of length 1, name of the column in data with the group ID.

group_reference

Atomic of length 1, element of the group column that indicates the control group. (The other groups may be treatment groups.)

patient

Character of length 1, name of the column in data with the patient ID.

m_mu

Numeric of length 1, prior mean of the mean mu in the simple model.

s_mu

Numeric of length 1, prior standard deviation of the mean mu in the simple model.

s_sigma

Numeric of length 1, uniform prior upper bound of the residual standard deviation sigma in the simple model.

m_omega_current

Numeric with length 1, m_omega value of the current study. Inserted as the final component of the m_omega column in the output.

s_omega_current

Numeric with length 1, s_omega value of the current study. Inserted as the final component of the s_omega column in the output.

n_chains

Number of MCMC chains to run.

n_adapt

Number of adaptation iterations to run.

n_warmup

Number of warmup iterations per chain to run.

n_iterations

Number of saved MCMC iterations per chain to run.

quiet

Logical of length 1, TRUE to suppress R console output.

Value

A tidy data frame of hyperparameter values for hb_mcmc_mixture(). The first several rows are for historical studies, and the last row is for the current study. Studies/rows are sorted in the order hb_mcmc_mixture() sorts them, so you can use columns m_omega

and s_omega for the same dataset and same values of other arguments directly in hb_mcmc_mixture().

Details

The model is a simple Bayesian model with a normal likelihood, an unknown mean mu, and an unknown standard deviation sigma. For each historical study, the posterior mean of mu becomes the corresponding component of m_omega in the output, and the posterior standard deviation of mu becomes the corresponding component of s_omega in the output. See the examples in this help file for a demonstration. m_omega and s_omega define the components of the mixture prior in hb_mcmc_mixture() that act as the contribution of the historical studies to the model.

See also

Examples

data_all_studies <- hb_sim_independent(n_continuous = 2)$data
data_all_studies$study <- paste0("study", data_all_studies$study)
hyperparameters <- hb_mcmc_mixture_hyperparameters(
  data = data_all_studies,
  response = "response",
  study = "study",
  study_reference = "study5",
  group = "group",
  group_reference = 1,
  patient = "patient",
  n_chains = 1,
  n_adapt = 100,
  n_warmup = 50,
  n_iterations = 50
)
print(hyperparameters)
#> # A tibble: 5 × 4
#>   study  study_index m_omega s_omega
#>   <chr>        <int>   <dbl>   <dbl>
#> 1 study1           1   1.03    0.109
#> 2 study2           2   0.506   0.129
#> 3 study3           3  -0.597   0.165
#> 4 study4           4  -1.69    0.100
#> 5 study5           5   0      30    
data_current_study <- dplyr::filter(data_all_studies, study == max(study))
hb_mcmc_mixture(
  data = data_current_study,
  response = "response",
  study = "study",
  study_reference = "study5",
  group = "group",
  group_reference = 1,
  patient = "patient",
  m_omega = hyperparameters$m_omega, # use hyperparams from historical data
  s_omega = hyperparameters$s_omega, # use hyperparams from historical data
  p_omega = rep(1 / nrow(hyperparameters), nrow(hyperparameters)),
  n_chains = 1,
  n_adapt = 100,
  n_warmup = 50,
  n_iterations = 50
)
#> # A tibble: 50 × 19
#>     alpha `beta[1]` `beta[2]` `delta[1]` delta…¹ omega…² omega…³ omega…⁴ omega…⁵
#>     <dbl>     <dbl>     <dbl>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 -1.09      0.506     0.737      0.743   0.518   0.940   0.302  -1.09    -1.80
#>  2 -1.01      0.467     0.661      0.701   0.490   1.04    0.569  -1.01    -1.75
#>  3 -1.14      0.467     0.758      0.773   0.537   1.07    0.459  -1.14    -1.72
#>  4 -1.01      0.469     0.730      0.784   0.483   1.10    0.599  -1.01    -1.68
#>  5 -0.983     0.509     0.748      0.701   0.432   0.765   0.493  -0.983   -1.59
#>  6 -0.998     0.446     0.766      0.652   0.486   1.03    0.549  -0.998   -1.63
#>  7 -1.06      0.494     0.734      0.725   0.546   0.984   0.579  -1.06    -1.57
#>  8 -0.977     0.509     0.781      0.832   0.464   1.01    0.351  -0.977   -1.81
#>  9 -0.967     0.463     0.727      0.760   0.480   0.961   0.607  -0.967   -1.58
#> 10 -1.01      0.465     0.697      0.806   0.447   1.28    0.524  -1.01    -1.64
#> # … with 40 more rows, 10 more variables: `omega[5]` <dbl>, `post_p[1]` <dbl>,
#> #   `post_p[2]` <dbl>, `post_p[3]` <dbl>, `post_p[4]` <dbl>, `post_p[5]` <dbl>,
#> #   sigma <dbl>, .chain <int>, .iteration <int>, .draw <int>, and abbreviated
#> #   variable names ¹​`delta[2]`, ²​`omega[1]`, ³​`omega[2]`, ⁴​`omega[3]`,
#> #   ⁵​`omega[4]`